app/requirements.txt
# requirements.txt
fastapi            # satisfies ADK >=0.115.0
starlette           # matches FastAPI 0.115.x
uvicorn[standard]
pydantic            # compatible with FastAPI 0.115.x
typing-extensions
anyio

# Vertex + ADK
google-adk
google-genai
google-cloud-aiplatform  # satisfies ADK >=1.87.0

# GCS + storage
google-cloud-storage

# App deps
faiss-cpu
numpy
python-dotenv
charset-normalizer
firebase-admin
